refactor: move auth middleware to dedicated folder
