+++ backend/src/middleware/authMiddleware.js
